# [Xamarin.Android.Build.Tasks] fixes for case-insensitive FS