--- a/build-tools/xa-prep-tasks/Xamarin.Android.BuildTools.PrepTasks/SystemUnzip.cs
+++ b/build-tools/xa-prep-tasks/Xamarin.Android.BuildTools.PrepTasks/SystemUnzip.cs
@@ -166,7 +166,6 @@ namespace Xamarin.Android.BuildTools.PrepTasks {
 					var destFileInfo = new FileInfo (dest);
 					if (!destFileInfo.Attributes.HasFlag (FileAttributes.ReparsePoint)) {
 						File.SetLastWriteTimeUtc (dest, DateTime.UtcNow);
-						File.SetLastAccessTimeUtc (dest, DateTime.UtcNow);
 					}
 				}
 			}