--- a/build-tools/xaprepare/xaprepare/Application/Utilities.cs
+++ b/build-tools/xaprepare/xaprepare/Application/Utilities.cs
@@ -582,7 +582,6 @@ namespace Xamarin.Android.Prepare
 					var destFileInfo = new FileInfo (filePath);
 					if (!destFileInfo.Attributes.HasFlag (FileAttributes.ReparsePoint)) {
 						File.SetLastWriteTimeUtc (filePath, stamp);
-						File.SetLastAccessTimeUtc (filePath, stamp);
 					}
 					break;
 				} catch (Exception e) {