--- a/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
@@ -441,7 +441,7 @@ namespace Xamarin.Android.Tasks
 			if (preserveTimestamp && File.Exists (path)) {
 				var timestamp = File.GetLastWriteTimeUtc (path);
 				File.WriteAllText (path, contents);
-				MonoAndroidHelper.SetLastAccessAndWriteTimeUtc (path, timestamp, Log);
+				File.SetLastWriteTimeUtc (path, timestamp);
 			} else {
 				File.WriteAllText (path, contents);
 			}